requirements.txt
aiopg
psycopg2-binary
docker
redis
aiohttp